Refactor AMD SMI component into separate modules
