+++ src/components/amd_smi/amds.c
@@ -1,4 +1,5 @@
 #include "amds.h"
+#include "amds_priv.h"
 #include "amdsmi.h"
 #include "htable.h"
 #include "papi.h"
@@ -9,120 +10,114 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <sys/stat.h>
-#define MAX_EVENTS_PER_DEVICE 1024
-#define AMDSMI_DISABLE_ESMI
+ #define MAX_EVENTS_PER_DEVICE 1024
 
 unsigned int _amd_smi_lock;
-typedef enum {
-  PAPI_MODE_READ = 1,
-  PAPI_MODE_WRITE,
-  PAPI_MODE_RDWR,
-} rocs_access_mode_e;
 /* Pointers to AMD SMI library functions (dynamically loaded) */
-static amdsmi_status_t (*amdsmi_init_p)(uint64_t);
-static amdsmi_status_t (*amdsmi_shut_down_p)(void);
-static amdsmi_status_t (*amdsmi_get_socket_handles_p)(uint32_t *, amdsmi_socket_handle *);
-static amdsmi_status_t (*amdsmi_get_processor_handles_by_type_p)(amdsmi_socket_handle, processor_type_t, amdsmi_processor_handle *,
-                                                                 uint32_t *);
-static amdsmi_status_t (*amdsmi_get_temp_metric_p)(amdsmi_processor_handle, amdsmi_temperature_type_t, amdsmi_temperature_metric_t,
-                                                   int64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_fan_rpms_p)(amdsmi_processor_handle, uint32_t, int64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_fan_speed_p)(amdsmi_processor_handle, uint32_t, int64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_fan_speed_max_p)(amdsmi_processor_handle, uint32_t, int64_t *);
-static amdsmi_status_t (*amdsmi_get_total_memory_p)(amdsmi_processor_handle, amdsmi_memory_type_t, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_memory_usage_p)(amdsmi_processor_handle, amdsmi_memory_type_t, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_activity_p)(amdsmi_processor_handle, amdsmi_engine_usage_t *);
-static amdsmi_status_t (*amdsmi_get_power_cap_info_p)(amdsmi_processor_handle, uint32_t, amdsmi_power_cap_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_power_cap_set_p)(amdsmi_processor_handle, uint32_t, uint64_t);
-static amdsmi_status_t (*amdsmi_get_power_info_p)(amdsmi_processor_handle, amdsmi_power_info_t *);
-static amdsmi_status_t (*amdsmi_set_power_cap_p)(amdsmi_processor_handle, uint32_t, uint64_t);
-static amdsmi_status_t (*amdsmi_get_gpu_pci_throughput_p)(amdsmi_processor_handle, uint64_t *, uint64_t *, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_pci_replay_counter_p)(amdsmi_processor_handle, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_clk_freq_p)(amdsmi_processor_handle, amdsmi_clk_type_t, amdsmi_frequencies_t *);
-static amdsmi_status_t (*amdsmi_set_clk_freq_p)(amdsmi_processor_handle, amdsmi_clk_type_t, uint64_t);
-static amdsmi_status_t (*amdsmi_get_gpu_metrics_info_p)(amdsmi_processor_handle, amdsmi_gpu_metrics_t *);
+amdsmi_status_t (*amdsmi_init_p)(uint64_t);
+amdsmi_status_t (*amdsmi_shut_down_p)(void);
+amdsmi_status_t (*amdsmi_get_socket_handles_p)(uint32_t *, amdsmi_socket_handle *);
+amdsmi_status_t (*amdsmi_get_processor_handles_by_type_p)(amdsmi_socket_handle, processor_type_t, amdsmi_processor_handle *,
+                                                          uint32_t *);
+amdsmi_status_t (*amdsmi_get_temp_metric_p)(amdsmi_processor_handle, amdsmi_temperature_type_t, amdsmi_temperature_metric_t,
+                                            int64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_fan_rpms_p)(amdsmi_processor_handle, uint32_t, int64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_fan_speed_p)(amdsmi_processor_handle, uint32_t, int64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_fan_speed_max_p)(amdsmi_processor_handle, uint32_t, int64_t *);
+amdsmi_status_t (*amdsmi_get_total_memory_p)(amdsmi_processor_handle, amdsmi_memory_type_t, uint64_t *);
+amdsmi_status_t (*amdsmi_get_memory_usage_p)(amdsmi_processor_handle, amdsmi_memory_type_t, uint64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_activity_p)(amdsmi_processor_handle, amdsmi_engine_usage_t *);
+amdsmi_status_t (*amdsmi_get_power_cap_info_p)(amdsmi_processor_handle, uint32_t, amdsmi_power_cap_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_power_cap_set_p)(amdsmi_processor_handle, uint32_t, uint64_t);
+amdsmi_status_t (*amdsmi_get_power_info_p)(amdsmi_processor_handle, amdsmi_power_info_t *);
+amdsmi_status_t (*amdsmi_set_power_cap_p)(amdsmi_processor_handle, uint32_t, uint64_t);
+amdsmi_status_t (*amdsmi_get_gpu_pci_throughput_p)(amdsmi_processor_handle, uint64_t *, uint64_t *, uint64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_pci_replay_counter_p)(amdsmi_processor_handle, uint64_t *);
+amdsmi_status_t (*amdsmi_get_clk_freq_p)(amdsmi_processor_handle, amdsmi_clk_type_t, amdsmi_frequencies_t *);
+amdsmi_status_t (*amdsmi_set_clk_freq_p)(amdsmi_processor_handle, amdsmi_clk_type_t, uint64_t);
+amdsmi_status_t (*amdsmi_get_gpu_metrics_info_p)(amdsmi_processor_handle, amdsmi_gpu_metrics_t *);
 
 /* Additional read-only AMD SMI function pointers */
-static amdsmi_status_t (*amdsmi_get_lib_version_p)(amdsmi_version_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_driver_info_p)(amdsmi_processor_handle, amdsmi_driver_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_asic_info_p)(amdsmi_processor_handle, amdsmi_asic_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_board_info_p)(amdsmi_processor_handle, amdsmi_board_info_t *);
-static amdsmi_status_t (*amdsmi_get_fw_info_p)(amdsmi_processor_handle, amdsmi_fw_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_vbios_info_p)(amdsmi_processor_handle, amdsmi_vbios_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_device_uuid_p)(amdsmi_processor_handle, unsigned int *, char *);
-static amdsmi_status_t (*amdsmi_get_gpu_enumeration_info_p)(amdsmi_processor_handle, amdsmi_enumeration_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_vendor_name_p)(amdsmi_processor_handle, char *, size_t);
-static amdsmi_status_t (*amdsmi_get_gpu_vram_vendor_p)(amdsmi_processor_handle, char *, uint32_t);
-static amdsmi_status_t (*amdsmi_get_gpu_subsystem_name_p)(amdsmi_processor_handle, char *, size_t);
-static amdsmi_status_t (*amdsmi_get_link_metrics_p)(amdsmi_processor_handle, amdsmi_link_metrics_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_process_list_p)(amdsmi_processor_handle, uint32_t *, amdsmi_proc_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_ecc_enabled_p)(amdsmi_processor_handle, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_total_ecc_count_p)(amdsmi_processor_handle, amdsmi_error_count_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_ecc_count_p)(amdsmi_processor_handle, amdsmi_gpu_block_t, amdsmi_error_count_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_ecc_status_p)(amdsmi_processor_handle, amdsmi_gpu_block_t, amdsmi_ras_err_state_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_compute_partition_p)(amdsmi_processor_handle, char *, uint32_t);
-static amdsmi_status_t (*amdsmi_get_gpu_memory_partition_p)(amdsmi_processor_handle, char *, uint32_t);
-static amdsmi_status_t (*amdsmi_get_gpu_accelerator_partition_profile_p)(amdsmi_processor_handle, amdsmi_accelerator_partition_profile_t *,
-                                                                         uint32_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_id_p)(amdsmi_processor_handle, uint16_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_revision_p)(amdsmi_processor_handle, uint16_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_subsystem_id_p)(amdsmi_processor_handle, uint16_t *);
+amdsmi_status_t (*amdsmi_get_lib_version_p)(amdsmi_version_t *);
+amdsmi_status_t (*amdsmi_get_gpu_driver_info_p)(amdsmi_processor_handle, amdsmi_driver_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_asic_info_p)(amdsmi_processor_handle, amdsmi_asic_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_board_info_p)(amdsmi_processor_handle, amdsmi_board_info_t *);
+amdsmi_status_t (*amdsmi_get_fw_info_p)(amdsmi_processor_handle, amdsmi_fw_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_vbios_info_p)(amdsmi_processor_handle, amdsmi_vbios_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_device_uuid_p)(amdsmi_processor_handle, unsigned int *, char *);
+amdsmi_status_t (*amdsmi_get_gpu_enumeration_info_p)(amdsmi_processor_handle, amdsmi_enumeration_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_vendor_name_p)(amdsmi_processor_handle, char *, size_t);
+amdsmi_status_t (*amdsmi_get_gpu_vram_vendor_p)(amdsmi_processor_handle, char *, uint32_t);
+amdsmi_status_t (*amdsmi_get_gpu_subsystem_name_p)(amdsmi_processor_handle, char *, size_t);
+amdsmi_status_t (*amdsmi_get_link_metrics_p)(amdsmi_processor_handle, amdsmi_link_metrics_t *);
+amdsmi_status_t (*amdsmi_get_gpu_process_list_p)(amdsmi_processor_handle, uint32_t *, amdsmi_proc_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_ecc_enabled_p)(amdsmi_processor_handle, uint64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_total_ecc_count_p)(amdsmi_processor_handle, amdsmi_error_count_t *);
+amdsmi_status_t (*amdsmi_get_gpu_ecc_count_p)(amdsmi_processor_handle, amdsmi_gpu_block_t, amdsmi_error_count_t *);
+amdsmi_status_t (*amdsmi_get_gpu_ecc_status_p)(amdsmi_processor_handle, amdsmi_gpu_block_t, amdsmi_ras_err_state_t *);
+amdsmi_status_t (*amdsmi_get_gpu_compute_partition_p)(amdsmi_processor_handle, char *, uint32_t);
+amdsmi_status_t (*amdsmi_get_gpu_memory_partition_p)(amdsmi_processor_handle, char *, uint32_t);
+amdsmi_status_t (*amdsmi_get_gpu_accelerator_partition_profile_p)(amdsmi_processor_handle, amdsmi_accelerator_partition_profile_t *,
+                                                                  uint32_t *);
+amdsmi_status_t (*amdsmi_get_gpu_id_p)(amdsmi_processor_handle, uint16_t *);
+amdsmi_status_t (*amdsmi_get_gpu_revision_p)(amdsmi_processor_handle, uint16_t *);
+amdsmi_status_t (*amdsmi_get_gpu_subsystem_id_p)(amdsmi_processor_handle, uint16_t *);
 // (*amdsmi_get_gpu_virtualization_mode_p)(amdsmi_processor_handle,amdsmi_virtualization_mode_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_pci_bandwidth_p)(amdsmi_processor_handle, amdsmi_pcie_bandwidth_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_bdf_id_p)(amdsmi_processor_handle, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_topo_numa_affinity_p)(amdsmi_processor_handle, int32_t *);
-static amdsmi_status_t (*amdsmi_get_energy_count_p)(amdsmi_processor_handle, uint64_t *, float *, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_power_profile_presets_p)(amdsmi_processor_handle, uint32_t, amdsmi_power_profile_status_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_cache_info_p)(amdsmi_processor_handle, amdsmi_gpu_cache_info_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_mem_overdrive_level_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_od_volt_curve_regions_p)(amdsmi_processor_handle, uint32_t *, amdsmi_freq_volt_region_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_od_volt_info_p)(amdsmi_processor_handle, amdsmi_od_volt_freq_data_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_overdrive_level_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_perf_level_p)(amdsmi_processor_handle, amdsmi_dev_perf_level_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_pm_metrics_info_p)(amdsmi_processor_handle, amdsmi_name_value_t **, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_ras_feature_info_p)(amdsmi_processor_handle, amdsmi_ras_feature_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_reg_table_info_p)(amdsmi_processor_handle, amdsmi_reg_type_t, amdsmi_name_value_t **, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_volt_metric_p)(amdsmi_processor_handle, amdsmi_voltage_type_t, amdsmi_voltage_metric_t, int64_t *);
-static amdsmi_status_t (*amdsmi_get_gpu_vram_info_p)(amdsmi_processor_handle, amdsmi_vram_info_t *);
-static amdsmi_status_t (*amdsmi_get_pcie_info_p)(amdsmi_processor_handle, amdsmi_pcie_info_t *);
-static amdsmi_status_t (*amdsmi_get_processor_count_from_handles_p)(amdsmi_processor_handle *, uint32_t *, uint32_t *, uint32_t *, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_soc_pstate_p)(amdsmi_processor_handle, amdsmi_dpm_policy_t *);
-static amdsmi_status_t (*amdsmi_get_xgmi_plpd_p)(amdsmi_processor_handle, amdsmi_dpm_policy_t *);
+amdsmi_status_t (*amdsmi_get_gpu_pci_bandwidth_p)(amdsmi_processor_handle, amdsmi_pcie_bandwidth_t *);
+amdsmi_status_t (*amdsmi_get_gpu_bdf_id_p)(amdsmi_processor_handle, uint64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_topo_numa_affinity_p)(amdsmi_processor_handle, int32_t *);
+amdsmi_status_t (*amdsmi_get_energy_count_p)(amdsmi_processor_handle, uint64_t *, float *, uint64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_power_profile_presets_p)(amdsmi_processor_handle, uint32_t, amdsmi_power_profile_status_t *);
+amdsmi_status_t (*amdsmi_get_gpu_cache_info_p)(amdsmi_processor_handle, amdsmi_gpu_cache_info_t *);
+amdsmi_status_t (*amdsmi_get_gpu_mem_overdrive_level_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_gpu_od_volt_curve_regions_p)(amdsmi_processor_handle, uint32_t *, amdsmi_freq_volt_region_t *);
+amdsmi_status_t (*amdsmi_get_gpu_od_volt_info_p)(amdsmi_processor_handle, amdsmi_od_volt_freq_data_t *);
+amdsmi_status_t (*amdsmi_get_gpu_overdrive_level_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_gpu_perf_level_p)(amdsmi_processor_handle, amdsmi_dev_perf_level_t *);
+amdsmi_status_t (*amdsmi_get_gpu_pm_metrics_info_p)(amdsmi_processor_handle, amdsmi_name_value_t **, uint32_t *);
+amdsmi_status_t (*amdsmi_get_gpu_ras_feature_info_p)(amdsmi_processor_handle, amdsmi_ras_feature_t *);
+amdsmi_status_t (*amdsmi_get_gpu_reg_table_info_p)(amdsmi_processor_handle, amdsmi_reg_type_t, amdsmi_name_value_t **, uint32_t *);
+amdsmi_status_t (*amdsmi_get_gpu_volt_metric_p)(amdsmi_processor_handle, amdsmi_voltage_type_t, amdsmi_voltage_metric_t, int64_t *);
+amdsmi_status_t (*amdsmi_get_gpu_vram_info_p)(amdsmi_processor_handle, amdsmi_vram_info_t *);
+amdsmi_status_t (*amdsmi_get_pcie_info_p)(amdsmi_processor_handle, amdsmi_pcie_info_t *);
+amdsmi_status_t (*amdsmi_get_processor_count_from_handles_p)(amdsmi_processor_handle *, uint32_t *, uint32_t *, uint32_t *, uint32_t *);
+amdsmi_status_t (*amdsmi_get_soc_pstate_p)(amdsmi_processor_handle, amdsmi_dpm_policy_t *);
+amdsmi_status_t (*amdsmi_get_xgmi_plpd_p)(amdsmi_processor_handle, amdsmi_dpm_policy_t *);
 
 #ifndef AMDSMI_DISABLE_ESMI
 /* CPU function pointers */
-static amdsmi_status_t (*amdsmi_get_cpu_handles_p)(uint32_t *, amdsmi_processor_handle *);
-static amdsmi_status_t (*amdsmi_get_cpucore_handles_p)(uint32_t *, amdsmi_processor_handle *);
-static amdsmi_status_t (*amdsmi_get_cpu_socket_power_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_socket_power_cap_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_socket_power_cap_max_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_core_energy_p)(amdsmi_processor_handle, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_socket_energy_p)(amdsmi_processor_handle, uint64_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_smu_fw_version_p)(amdsmi_processor_handle, amdsmi_smu_fw_version_t *);
-static amdsmi_status_t (*amdsmi_get_threads_per_core_p)(uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_family_p)(uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_model_p)(uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_core_boostlimit_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_socket_current_active_freq_limit_p)(amdsmi_processor_handle, uint16_t *, char **);
-static amdsmi_status_t (*amdsmi_get_cpu_socket_freq_range_p)(amdsmi_processor_handle, uint16_t *, uint16_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_core_current_freq_limit_p)(amdsmi_processor_handle, uint32_t *);
-static amdsmi_status_t (*amdsmi_get_minmax_bandwidth_between_processors_p)(amdsmi_processor_handle, amdsmi_processor_handle, uint64_t *,
-                                                                           uint64_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_dimm_temp_range_and_refresh_rate_p)(amdsmi_processor_handle, uint8_t,
-                                                                            amdsmi_temp_range_refresh_rate_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_dimm_power_consumption_p)(amdsmi_processor_handle, uint8_t, amdsmi_dimm_power_t *);
-static amdsmi_status_t (*amdsmi_get_cpu_dimm_thermal_sensor_p)(amdsmi_processor_handle, uint8_t, amdsmi_dimm_thermal_t *);
+amdsmi_status_t (*amdsmi_get_cpu_handles_p)(uint32_t *, amdsmi_processor_handle *);
+amdsmi_status_t (*amdsmi_get_cpucore_handles_p)(uint32_t *, amdsmi_processor_handle *);
+amdsmi_status_t (*amdsmi_get_cpu_socket_power_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_socket_power_cap_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_socket_power_cap_max_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_core_energy_p)(amdsmi_processor_handle, uint64_t *);
+amdsmi_status_t (*amdsmi_get_cpu_socket_energy_p)(amdsmi_processor_handle, uint64_t *);
+amdsmi_status_t (*amdsmi_get_cpu_smu_fw_version_p)(amdsmi_processor_handle, amdsmi_smu_fw_version_t *);
+amdsmi_status_t (*amdsmi_get_threads_per_core_p)(uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_family_p)(uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_model_p)(uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_core_boostlimit_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_cpu_socket_current_active_freq_limit_p)(amdsmi_processor_handle, uint16_t *, char **);
+amdsmi_status_t (*amdsmi_get_cpu_socket_freq_range_p)(amdsmi_processor_handle, uint16_t *, uint16_t *);
+amdsmi_status_t (*amdsmi_get_cpu_core_current_freq_limit_p)(amdsmi_processor_handle, uint32_t *);
+amdsmi_status_t (*amdsmi_get_minmax_bandwidth_between_processors_p)(amdsmi_processor_handle, amdsmi_processor_handle, uint64_t *,
+                                                                    uint64_t *);
+amdsmi_status_t (*amdsmi_get_cpu_dimm_temp_range_and_refresh_rate_p)(amdsmi_processor_handle, uint8_t,
+                                                                     amdsmi_temp_range_refresh_rate_t *);
+amdsmi_status_t (*amdsmi_get_cpu_dimm_power_consumption_p)(amdsmi_processor_handle, uint8_t, amdsmi_dimm_power_t *);
+amdsmi_status_t (*amdsmi_get_cpu_dimm_thermal_sensor_p)(amdsmi_processor_handle, uint8_t, amdsmi_dimm_thermal_t *);
 #endif
 /* Global device list and count */
-static int32_t device_count = 0;
-static amdsmi_processor_handle *device_handles = NULL;
-static int32_t device_mask = 0;
-static int32_t gpu_count = 0;
-static int32_t cpu_count = 0;
-static amdsmi_processor_handle **cpu_core_handles = NULL;
-static uint32_t *cores_per_socket = NULL;
+int32_t device_count = 0;
+amdsmi_processor_handle *device_handles = NULL;
+int32_t device_mask = 0;
+int32_t gpu_count = 0;
+int32_t cpu_count = 0;
+amdsmi_processor_handle **cpu_core_handles = NULL;
+uint32_t *cores_per_socket = NULL;
 static void *amds_dlp = NULL;
-static void *htable = NULL;
+void *htable = NULL;
 static char error_string[PAPI_MAX_STR_LEN + 1];
 /* forward declarations for internal helpers */
 static int load_amdsmi_sym(void);
@@ -130,54 +125,8 @@ static int init_device_table(void);
 static int shutdown_device_table(void);
 static int init_event_table(void);
 static int shutdown_event_table(void);
-/* Event descriptor structure for native events */
-typedef struct native_event {
-  unsigned int id;
-  char *name;
-  char *descr;
-  int32_t device;                           /* device index or -1 if not applicable */
-  uint64_t value;                           /* last read value or set value */
-  uint32_t mode;                            /* access mode (read/write) */
-  uint32_t variant;                         /* variant index (for metric type, etc.) */
-  uint32_t subvariant;                      /* subvariant index (for sensor index or sub-type) */
-  int (*open_func)(struct native_event *);  /* optional open (reserve resources) */
-  int (*close_func)(struct native_event *); /* optional close (release resources) */
-  int (*start_func)(struct native_event *); /* optional start (begin counting) */
-  int (*stop_func)(struct native_event *);  /* optional stop (stop counting) */
-  int (*access_func)(int mode, void *arg);  /* read or write the event value */
-} native_event_t;
-/* Table of all native events */
-typedef struct {
-  native_event_t *events;
-  int count;
-} native_event_table_t;
 static native_event_table_t ntv_table;
-static native_event_table_t *ntv_table_p = NULL;
-/* Locking device usage for contexts */
-static int acquire_devices(unsigned int *events_id, int num_events, int32_t *bitmask) {
-  int32_t mask_acq = 0;
-  for (int i = 0; i < num_events; ++i) {
-    int32_t dev_id = ntv_table_p->events[events_id[i]].device;
-    if (dev_id < 0)
-      continue;
-    mask_acq |= (1 << dev_id);
-  }
-  if (mask_acq & device_mask) {
-    return PAPI_ECNFLCT; // conflict: device already in use
-  }
-  device_mask |= mask_acq;
-  *bitmask = mask_acq;
-  return PAPI_OK;
-}
-static int release_devices(int32_t *bitmask) {
-  int32_t mask_rel = *bitmask;
-  if ((mask_rel & device_mask) != mask_rel) {
-    return PAPI_EMISC;
-  }
-  device_mask ^= mask_rel;
-  *bitmask = 0;
-  return PAPI_OK;
-}
+native_event_table_t *ntv_table_p = NULL;
 
 /* Redirects stderr to /dev/null, returns dup of old stderr (or -1 on failure). */
 static int silence_stderr_begin(void) {
@@ -198,71 +147,6 @@ static void silence_stderr_end(int saved_fd) {
   }
 }
 
-/* Prototypes for event access (read/write) functions */
-static int access_amdsmi_temp_metric(int mode, void *arg);
-static int access_amdsmi_fan_speed(int mode, void *arg);
-static int access_amdsmi_fan_rpms(int mode, void *arg);
-static int access_amdsmi_mem_total(int mode, void *arg);
-static int access_amdsmi_mem_usage(int mode, void *arg);
-static int access_amdsmi_power_cap(int mode, void *arg);
-static int access_amdsmi_power_cap_range(int mode, void *arg);
-static int access_amdsmi_power_average(int mode, void *arg);
-static int access_amdsmi_pci_throughput(int mode, void *arg);
-static int access_amdsmi_pci_replay_counter(int mode, void *arg);
-static int access_amdsmi_clk_freq(int mode, void *arg);
-static int access_amdsmi_gpu_metrics(int mode, void *arg);
-static int access_amdsmi_gpu_info(int mode, void *arg);
-static int access_amdsmi_gpu_activity(int mode, void *arg);
-static int access_amdsmi_fan_speed_max(int mode, void *arg);
-static int access_amdsmi_pci_bandwidth(int mode, void *arg);
-static int access_amdsmi_energy_count(int mode, void *arg);
-static int access_amdsmi_power_profile_status(int mode, void *arg);
-static uint64_t _str_to_u64_hash(const char *s);
-static int access_amdsmi_uuid_hash(int mode, void *arg);
-static int access_amdsmi_gpu_string_hash(int mode, void *arg);
-static int access_amdsmi_enumeration_info(int mode, void *arg);
-static int access_amdsmi_asic_info(int mode, void *arg);
-static int access_amdsmi_link_metrics(int mode, void *arg);
-static int access_amdsmi_process_count(int mode, void *arg);
-static int access_amdsmi_ecc_total(int mode, void *arg);
-static int access_amdsmi_ecc_enabled_mask(int mode, void *arg);
-static int access_amdsmi_compute_partition_hash(int mode, void *arg);
-static int access_amdsmi_memory_partition_hash(int mode, void *arg);
-static int access_amdsmi_accelerator_num_partitions(int mode, void *arg);
-static int access_amdsmi_lib_version(int mode, void *arg);
-
-static int access_amdsmi_cache_size(int mode, void *arg);
-static int access_amdsmi_overdrive_level(int mode, void *arg);
-static int access_amdsmi_mem_overdrive_level(int mode, void *arg);
-static int access_amdsmi_od_volt_regions_count(int mode, void *arg);
-static int access_amdsmi_perf_level(int mode, void *arg);
-static int access_amdsmi_pm_metrics_count(int mode, void *arg);
-static int access_amdsmi_ras_ecc_schema(int mode, void *arg);
-static int access_amdsmi_reg_count(int mode, void *arg);
-static int access_amdsmi_voltage(int mode, void *arg);
-static int access_amdsmi_vram_width(int mode, void *arg);
-static int access_amdsmi_soc_pstate_id(int mode, void *arg);
-static int access_amdsmi_xgmi_plpd_id(int mode, void *arg);
-
-
-#ifndef AMDSMI_DISABLE_ESMI
-static int access_amdsmi_cpu_socket_power(int mode, void *arg);
-static int access_amdsmi_cpu_socket_energy(int mode, void *arg);
-static int access_amdsmi_cpu_socket_freq_limit(int mode, void *arg);
-static int access_amdsmi_cpu_socket_freq_range(int mode, void *arg);
-static int access_amdsmi_cpu_power_cap(int mode, void *arg);
-static int access_amdsmi_cpu_core_energy(int mode, void *arg);
-static int access_amdsmi_cpu_core_freq_limit(int mode, void *arg);
-static int access_amdsmi_cpu_core_boostlimit(int mode, void *arg);
-static int access_amdsmi_dimm_temp(int mode, void *arg);
-static int access_amdsmi_dimm_power(int mode, void *arg);
-static int access_amdsmi_dimm_range_refresh(int mode, void *arg);
-static int access_amdsmi_threads_per_core(int mode, void *arg);
-static int access_amdsmi_cpu_family(int mode, void *arg);
-static int access_amdsmi_cpu_model(int mode, void *arg);
-static int access_amdsmi_smu_fw_version(int mode, void *arg);
-static int access_amdsmi_xgmi_bandwidth(int mode, void *arg);
-#endif
 /* Simple open/close/start/stop functions (no special handling needed for most
  * events) */
 static int open_simple(native_event_t *event) {
@@ -598,132 +482,6 @@ int amds_shutdown(void) {
 int amds_err_get_last(const char **err_string) {
   if (err_string)
     *err_string = error_string;
-  return PAPI_OK;
-} 
-/* Event enumeration: iterate over native events */
-int amds_evt_enum(unsigned int *EventCode, int modifier) {
-  if (modifier == PAPI_ENUM_FIRST) {
-    if (ntv_table_p->count == 0) {
-      return PAPI_ENOEVNT;
-    }
-    *EventCode = 0;
-    return PAPI_OK;
-  } else if (modifier == PAPI_ENUM_EVENTS) {
-    if (*EventCode + 1 < (unsigned int)ntv_table_p->count) {
-      *EventCode = *EventCode + 1;
-      return PAPI_OK;
-    } else {
-      return PAPI_ENOEVNT;
-    }
-  }
-  return PAPI_EINVAL;
-}
-int amds_evt_code_to_name(unsigned int EventCode, char *name, int len) {
-  if (EventCode >= (unsigned int)ntv_table_p->count) {
-    return PAPI_EINVAL;
-  }
-  strncpy(name, ntv_table_p->events[EventCode].name, len);
-  return PAPI_OK;
-}
-int amds_evt_name_to_code(const char *name, unsigned int *EventCode) {
-  native_event_t *event = NULL;
-  int hret = htable_find(htable, name, (void **)&event);
-  if (hret != HTABLE_SUCCESS) {
-    return (hret == HTABLE_ENOVAL) ? PAPI_ENOEVNT : PAPI_ECMP;
-  }
-  *EventCode = event->id;
-  return PAPI_OK;
-}
-int amds_evt_code_to_descr(unsigned int EventCode, char *descr, int len) {
-  if (EventCode >= (unsigned int)ntv_table_p->count) {
-    return PAPI_EINVAL;
-  }
-  strncpy(descr, ntv_table_p->events[EventCode].descr, len);
-  return PAPI_OK;
-}
-/* Context management: open/close, start/stop, read/write, reset */
-struct amds_ctx {
-  int state;
-  unsigned int *events_id;
-  int num_events;
-  long long *counters;
-  int32_t device_mask;
-};
-int amds_ctx_open(unsigned int *event_ids, int num_events, amds_ctx_t *ctx) {
-  amds_ctx_t new_ctx = (amds_ctx_t)papi_calloc(1, sizeof(struct amds_ctx));
-  if (new_ctx == NULL) {
-    return PAPI_ENOMEM;
-  }
-  new_ctx->events_id = event_ids; // Store pointer (original approach)
-  new_ctx->num_events = num_events;
-  new_ctx->counters = (long long *)papi_calloc(num_events, sizeof(long long));
-  if (new_ctx->counters == NULL) {
-    papi_free(new_ctx);
-    return PAPI_ENOMEM;
-  }
-  // Acquire devices needed by these events to avoid conflicts
-  int papi_errno = acquire_devices(event_ids, num_events, &new_ctx->device_mask);
-  if (papi_errno != PAPI_OK) {
-    papi_free(new_ctx->counters);
-    papi_free(new_ctx);
-    return papi_errno;
-  }
-  *ctx = new_ctx;
-  return PAPI_OK;
-}
-int amds_ctx_close(amds_ctx_t ctx) {
-  if (!ctx)
-    return PAPI_OK;
-  // release device usage
-  release_devices(&ctx->device_mask);
-  papi_free(ctx->counters);
-  papi_free(ctx);
-  return PAPI_OK;
-}
-int amds_ctx_start(amds_ctx_t ctx) {
-  // No additional actions needed to start in this design (all reads are
-  // on-demand)
-  ctx->state |= AMDS_EVENTS_RUNNING;
-  return PAPI_OK;
-}
-int amds_ctx_stop(amds_ctx_t ctx) {
-  if (!(ctx->state & AMDS_EVENTS_RUNNING)) {
-    return PAPI_OK;
-  }
-  ctx->state &= ~AMDS_EVENTS_RUNNING;
-  return PAPI_OK;
-}
-int amds_ctx_read(amds_ctx_t ctx, long long **counts) {
-  int papi_errno = PAPI_OK;
-  for (int i = 0; i < ctx->num_events; ++i) {
-    unsigned int id = ctx->events_id[i];
-    papi_errno = ntv_table_p->events[id].access_func(PAPI_MODE_READ, &ntv_table_p->events[id]);
-    if (papi_errno != PAPI_OK) {
-      return papi_errno;
-    }
-    ctx->counters[i] = (long long)ntv_table_p->events[id].value;
-  }
-  *counts = ctx->counters;
-  return papi_errno;
-}
-int amds_ctx_write(amds_ctx_t ctx, long long *counts) {
-  int papi_errno = PAPI_OK;
-  for (int i = 0; i < ctx->num_events; ++i) {
-    unsigned int id = ctx->events_id[i];
-    ntv_table_p->events[id].value = counts[i];
-    papi_errno = ntv_table_p->events[id].access_func(PAPI_MODE_WRITE, &ntv_table_p->events[id]);
-    if (papi_errno != PAPI_OK) {
-      return papi_errno;
-    }
-  }
-  return papi_errno;
-}
-int amds_ctx_reset(amds_ctx_t ctx) {
-  for (int i = 0; i < ctx->num_events; ++i) {
-    unsigned int id = ctx->events_id[i];
-    ntv_table_p->events[id].value = 0;
-    ctx->counters[i] = 0;
-  }
   return PAPI_OK;
 }
 /* Initialize native event table: enumerate all supported events */
@@ -3103,336 +2861,6 @@ static int init_event_table(void) {
   return PAPI_OK;
 }
 
-/* -------- Helpers and new accessors (GPU read-only additions) -------- */
-static uint64_t _str_to_u64_hash(const char *s) {
-  /* djb2 64-bit */
-  uint64_t hash = 5381;
-  if (!s)
-    return 0;
-  int c;
-  while ((c = *s++)) {
-    hash = ((hash << 5) + hash) + (uint8_t)c;
-  }
-  return hash;
-}
-static int access_amdsmi_lib_version(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_lib_version_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  amdsmi_version_t vinfo;
-  amdsmi_status_t st = amdsmi_get_lib_version_p(&vinfo);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  switch (event->variant) {
-  case 0:
-    event->value = (int64_t)vinfo.major;
-    break;
-  case 1:
-    event->value = (int64_t)vinfo.minor;
-    break;
-  case 2:
-    event->value = (int64_t)vinfo.release;
-    break;
-  default:
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_uuid_hash(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_device_uuid_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  char buf[128] = {0};
-  unsigned int len = sizeof(buf);
-  amdsmi_status_t st = amdsmi_get_gpu_device_uuid_p(device_handles[event->device], &len, buf);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  event->value = (int64_t)_str_to_u64_hash(buf);
-  return PAPI_OK;
-}
-static int access_amdsmi_gpu_string_hash(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  char buf[256] = {0};
-  amdsmi_status_t st = AMDSMI_STATUS_NOT_SUPPORTED;
-  switch (event->variant) {
-  case 0: /* vendor name */
-    if (!amdsmi_get_gpu_vendor_name_p)
-      return PAPI_ENOSUPP;
-    st = amdsmi_get_gpu_vendor_name_p(device_handles[event->device], buf, sizeof(buf));
-    break;
-  case 1: /* vram vendor */
-    if (!amdsmi_get_gpu_vram_vendor_p)
-      return PAPI_ENOSUPP;
-    st = amdsmi_get_gpu_vram_vendor_p(device_handles[event->device], buf, sizeof(buf));
-    break;
-  case 2: /* subsystem name */
-    if (!amdsmi_get_gpu_subsystem_name_p)
-      return PAPI_ENOSUPP;
-    st = amdsmi_get_gpu_subsystem_name_p(device_handles[event->device], buf, sizeof(buf));
-    break;
-  case 3: /* driver name */
-  case 4: /* driver date */
-    if (!amdsmi_get_gpu_driver_info_p)
-      return PAPI_ENOSUPP;
-    {
-      amdsmi_driver_info_t dinfo;
-      st = amdsmi_get_gpu_driver_info_p(device_handles[event->device], &dinfo);
-      if (st == AMDSMI_STATUS_SUCCESS) {
-        if (event->variant == 3)
-          strncpy(buf, dinfo.driver_name, sizeof(buf) - 1);
-        else
-          strncpy(buf, dinfo.driver_date, sizeof(buf) - 1);
-      }
-    }
-    break;
-  case 5: /* vbios version */
-  case 6: /* vbios part number */
-  case 7: /* vbios build date */
-    if (!amdsmi_get_gpu_vbios_info_p)
-      return PAPI_ENOSUPP;
-    {
-      amdsmi_vbios_info_t vb;
-      st = amdsmi_get_gpu_vbios_info_p(device_handles[event->device], &vb);
-      if (st == AMDSMI_STATUS_SUCCESS) {
-        if (event->variant == 5)
-          strncpy(buf, vb.version, sizeof(buf) - 1);
-        else if (event->variant == 6)
-          strncpy(buf, vb.part_number, sizeof(buf) - 1);
-        else
-          strncpy(buf, vb.build_date, sizeof(buf) - 1);
-      }
-    }
-    break;
-  default:
-    return PAPI_ENOSUPP;
-  }
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  event->value = (int64_t)_str_to_u64_hash(buf);
-  return PAPI_OK;
-}
-static int access_amdsmi_enumeration_info(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_enumeration_info_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  amdsmi_enumeration_info_t info;
-  amdsmi_status_t st = amdsmi_get_gpu_enumeration_info_p(device_handles[event->device], &info);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  switch (event->variant) {
-  case 0:
-    event->value = (int64_t)info.drm_render;
-    break;
-  case 1:
-    event->value = (int64_t)info.drm_card;
-    break;
-  case 2:
-    event->value = (int64_t)info.hsa_id;
-    break;
-  case 3:
-    event->value = (int64_t)info.hip_id;
-    break;
-  default:
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_asic_info(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_asic_info_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  amdsmi_asic_info_t info;
-  amdsmi_status_t st = amdsmi_get_gpu_asic_info_p(device_handles[event->device], &info);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  switch (event->variant) {
-  case 0:
-    event->value = (int64_t)info.vendor_id;
-    break;
-  case 1:
-    event->value = (int64_t)info.device_id;
-    break;
-  case 2:
-    event->value = (int64_t)info.subvendor_id;
-    break;
-  case 3:
-    event->value = (int64_t)0 /* not provided in amdsmi_asic_info_t */;
-    break;
-  case 4:
-    event->value = (int64_t)info.rev_id;
-    break;
-  case 5:
-    event->value = (int64_t)info.num_of_compute_units;
-    break;
-  default:
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_link_metrics(int mode, void *arg) {
-    if (mode != PAPI_MODE_READ || !amdsmi_get_link_metrics_p) return PAPI_ENOSUPP;
-    native_event_t *event = (native_event_t *)arg;
-    if (event->device < 0 || event->device >= device_count || !device_handles[event->device]) return PAPI_EMISC;
-
-    amdsmi_link_metrics_t lm; memset(&lm, 0, sizeof(lm));
-    if (amdsmi_get_link_metrics_p(device_handles[event->device], &lm) != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-
-    uint64_t acc = 0;
-    uint32_t n = lm.num_links;
-    if (n > AMDSMI_MAX_NUM_XGMI_PHYSICAL_LINK) n = AMDSMI_MAX_NUM_XGMI_PHYSICAL_LINK;
-
-    for (uint32_t i = 0; i < n; ++i) {
-        if (event->subvariant != 0 && lm.links[i].link_type != (uint32_t)event->subvariant) continue;
-        switch (event->variant) {
-            case 0: acc += lm.links[i].read;         break;  /* KB */
-            case 1: acc += lm.links[i].write;        break;  /* KB */
-            case 2: acc += lm.links[i].bit_rate;     break;  /* Gb/s */
-            case 3: acc += lm.links[i].max_bandwidth;break;  /* Gb/s */
-            default: return PAPI_ENOSUPP;
-        }
-    }
-
-    if (acc > (uint64_t)INT64_MAX) acc = (uint64_t)INT64_MAX; /* defensive */
-    event->value = (int64_t)acc;
-    return PAPI_OK;
-}
-
-static int access_amdsmi_process_count(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_process_list_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  uint32_t cap = 32;
-  int tries = 0;
-  amdsmi_proc_info_t *list = NULL;
-  amdsmi_status_t st;
-  while (tries++ < 4) {
-    list = (amdsmi_proc_info_t *)papi_calloc(cap, sizeof(amdsmi_proc_info_t));
-    if (!list)
-      return PAPI_ENOMEM;
-    uint32_t maxp = cap;
-    st = amdsmi_get_gpu_process_list_p(device_handles[event->device], &maxp, list);
-    if (st == AMDSMI_STATUS_OUT_OF_RESOURCES) {
-      papi_free(list);
-      cap *= 2;
-      continue;
-    }
-    if (st != AMDSMI_STATUS_SUCCESS) {
-      papi_free(list);
-      return PAPI_EMISC;
-    }
-    event->value = (int64_t)maxp;
-    papi_free(list);
-    return PAPI_OK;
-  }
-  return PAPI_EMISC;
-}
-static int access_amdsmi_ecc_total(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_total_ecc_count_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-
-  amdsmi_error_count_t ec;
-  memset(&ec, 0, sizeof(ec)); // <-- important
-  amdsmi_status_t st = amdsmi_get_gpu_total_ecc_count_p(device_handles[event->device], &ec);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-
-  uint64_t v = (event->variant == 0)   ? ec.correctable_count
-               : (event->variant == 1) ? ec.uncorrectable_count
-                                       : ec.deferred_count; // some ASICs may not set this
-  if ((int64_t)v < 0)
-    return PAPI_ENOSUPP; // defensive: treat bogus as unsupported
-  event->value = (v > (uint64_t)INT64_MAX) ? INT64_MAX : (int64_t)v;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_ecc_enabled_mask(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_ecc_enabled_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  uint64_t mask = 0;
-  amdsmi_status_t st = amdsmi_get_gpu_ecc_enabled_p(device_handles[event->device], &mask);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  event->value = (int64_t)mask;
-  return PAPI_OK;
-}
-static int access_amdsmi_compute_partition_hash(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_compute_partition_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  char buf[128] = {0};
-  amdsmi_status_t st = amdsmi_get_gpu_compute_partition_p(device_handles[event->device], buf, sizeof(buf));
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  event->value = (int64_t)_str_to_u64_hash(buf);
-  return PAPI_OK;
-}
-static int access_amdsmi_memory_partition_hash(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_memory_partition_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  char buf[128] = {0};
-  amdsmi_status_t st = amdsmi_get_gpu_memory_partition_p(device_handles[event->device], buf, sizeof(buf));
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  event->value = (int64_t)_str_to_u64_hash(buf);
-  return PAPI_OK;
-}
-static int access_amdsmi_accelerator_num_partitions(int mode, void *arg) {
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_accelerator_partition_profile_p)
-    return PAPI_ENOSUPP;
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles[event->device])
-    return PAPI_EMISC;
-  amdsmi_accelerator_partition_profile_t prof;
-  uint32_t ids[AMDSMI_MAX_ACCELERATOR_PARTITIONS] = {0};
-  amdsmi_status_t st = amdsmi_get_gpu_accelerator_partition_profile_p(device_handles[event->device], &prof, ids);
-  if (st != AMDSMI_STATUS_SUCCESS)
-    return PAPI_EMISC;
-  event->value = (int64_t)prof.num_partitions;
-  return PAPI_OK;
-}
 static int shutdown_event_table(void) {
   // Remove all events from hash table and free their names/descr
   for (int i = 0; i < ntv_table.count; ++i) {
@@ -3469,904 +2897,5 @@ static int shutdown_device_table(void) {
   device_count = 0;
   gpu_count = 0;
   cpu_count = 0;
-  return PAPI_OK;
-}
-/* Access function implementations (read/write operations for each event) */
-static int access_amdsmi_temp_metric(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC; /* ensure device handle is valid */
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_status_t status = amdsmi_get_temp_metric_p(device_handles[event->device], (amdsmi_temperature_type_t)event->subvariant,
-                                                    (amdsmi_temperature_metric_t)event->variant, (int64_t *)&event->value);
-  return (status == AMDSMI_STATUS_SUCCESS ? PAPI_OK : PAPI_EMISC);
-}
-static int access_amdsmi_fan_rpms(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  int64_t speed = 0;
-  amdsmi_status_t status = amdsmi_get_gpu_fan_rpms_p(device_handles[event->device], event->subvariant, &speed);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = speed;
-  return PAPI_OK;
-}
-static int access_amdsmi_fan_speed(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP; // writing fan speed not supported
-  }
-  int64_t val = 0;
-  amdsmi_status_t status = amdsmi_get_gpu_fan_speed_p(device_handles[event->device], event->subvariant, &val);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = val;
-  return PAPI_OK;
-}
-static int access_amdsmi_mem_total(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  uint64_t data = 0;
-  amdsmi_status_t status = amdsmi_get_total_memory_p(device_handles[event->device], (amdsmi_memory_type_t)event->variant, &data);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)data;
-  return PAPI_OK;
-}
-static int access_amdsmi_mem_usage(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  uint64_t data = 0;
-  amdsmi_status_t status = amdsmi_get_memory_usage_p(device_handles[event->device], (amdsmi_memory_type_t)event->variant, &data);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)data;
-  return PAPI_OK;
-}
-static int access_amdsmi_power_cap(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode == PAPI_MODE_READ) {
-    // Read current power cap
-    amdsmi_power_cap_info_t info;
-    amdsmi_status_t status = amdsmi_get_power_cap_info_p(device_handles[event->device], 0, &info); // sensor index 0
-    if (status != AMDSMI_STATUS_SUCCESS) {
-      return PAPI_EMISC;
-    }
-    event->value = (int64_t)info.power_cap;
-    return PAPI_OK;
-  } else if (mode == PAPI_MODE_WRITE) {
-    // Set new power cap (value expected in microWatts if API uses uW)
-    uint64_t new_cap = (uint64_t)event->value;
-    amdsmi_status_t status = amdsmi_set_power_cap_p(device_handles[event->device], 0, new_cap);
-    return (status == AMDSMI_STATUS_SUCCESS ? PAPI_OK : PAPI_EMISC);
-  }
-  return PAPI_ENOSUPP;
-}
-static int access_amdsmi_power_cap_range(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  amdsmi_power_cap_info_t info;
-  amdsmi_status_t status = amdsmi_get_power_cap_info_p(device_handles[event->device], 0, &info);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  if (event->variant == 1) {
-    event->value = (int64_t)info.min_power_cap;
-  } else if (event->variant == 2) {
-    event->value = (int64_t)info.max_power_cap;
-  } else {
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_power_average(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  amdsmi_power_info_t power;
-  amdsmi_status_t status = amdsmi_get_power_info_p(device_handles[event->device], &power);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)power.average_socket_power;
-  return PAPI_OK;
-}
-static int access_amdsmi_pci_throughput(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  uint64_t sent = 0, received = 0, max_pkt = 0;
-  amdsmi_status_t status = amdsmi_get_gpu_pci_throughput_p(device_handles[event->device], &sent, &received, &max_pkt);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  switch (event->variant) {
-  case 0:
-    event->value = (int64_t)sent;
-    break;
-  case 1:
-    event->value = (int64_t)received;
-    break;
-  case 2:
-    event->value = (int64_t)max_pkt;
-    break;
-  default:
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_pci_replay_counter(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  uint64_t counter = 0;
-  amdsmi_status_t status = amdsmi_get_gpu_pci_replay_counter_p(device_handles[event->device], &counter);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)counter;
-  return PAPI_OK;
-}
-static int access_amdsmi_clk_freq(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  amdsmi_frequencies_t freq_info;
-  amdsmi_status_t status = amdsmi_get_clk_freq_p(device_handles[event->device], AMDSMI_CLK_TYPE_SYS, &freq_info);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  // Simplified: variant 0 -> count, 1 -> current frequency, >=2 -> specific
-  // index
-  if (event->subvariant == 0) {
-    event->value = freq_info.num_supported;
-  } else if (event->subvariant == 1) {
-    if (freq_info.num_supported > 0) {
-      event->value = freq_info.frequency[0]; // assume first is current
-    } else {
-      event->value = 0;
-    }
-  } else {
-    int idx = event->subvariant - 2;
-    if (idx >= 0 && idx < (int)freq_info.num_supported) {
-      event->value = freq_info.frequency[idx];
-    } else {
-      return PAPI_EMISC;
-    }
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_gpu_metrics(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ)
-    return PAPI_ENOSUPP;
-  amdsmi_gpu_metrics_t metrics;
-  amdsmi_status_t status = amdsmi_get_gpu_metrics_info_p(device_handles[event->device], &metrics);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  // Parsing of metrics is not fully implemented; just return OK.
-  // (In a full implementation, event->variant or subvariant would select a
-  // specific field of 'metrics'.)
-  return PAPI_OK;
-}
-static int access_amdsmi_gpu_info(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_status_t status;
-  switch (event->variant) {
-  case 0: {
-    uint16_t id = 0;
-    status = amdsmi_get_gpu_id_p(device_handles[event->device], &id);
-    if (status == AMDSMI_STATUS_SUCCESS) {
-      event->value = id;
-    }
-    break;
-  }
-  case 1: {
-    uint16_t rev = 0;
-    status = amdsmi_get_gpu_revision_p(device_handles[event->device], &rev);
-    if (status == AMDSMI_STATUS_SUCCESS) {
-      event->value = rev;
-    }
-    break;
-  }
-  case 2: {
-    uint16_t subid = 0;
-    status = amdsmi_get_gpu_subsystem_id_p(device_handles[event->device], &subid);
-    if (status == AMDSMI_STATUS_SUCCESS) {
-      event->value = subid;
-    }
-    break;
-  }
-  case 3: {
-    uint64_t bdfid = 0;
-    status = amdsmi_get_gpu_bdf_id_p(device_handles[event->device], &bdfid);
-    if (status == AMDSMI_STATUS_SUCCESS) {
-      event->value = (int64_t)bdfid;
-    }
-    break;
-  }
-  /*case 4: {
-      amdsmi_virtualization_mode_t mode_val;
-      status =
-  amdsmi_get_gpu_virtualization_mode_p(device_handles[event->device],
-  &mode_val); if (status == AMDSMI_STATUS_SUCCESS) { event->value = mode_val;
-      }
-      break;
-  }*/
-  case 5: {
-    int32_t numa_node = -1;
-    status = amdsmi_get_gpu_topo_numa_affinity_p(device_handles[event->device], &numa_node);
-    if (status == AMDSMI_STATUS_SUCCESS) {
-      event->value = numa_node;
-    }
-    break;
-  }
-  default:
-    return PAPI_EMISC;
-  }
-  return (status == AMDSMI_STATUS_SUCCESS ? PAPI_OK : PAPI_EMISC);
-}
-static int access_amdsmi_gpu_activity(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_engine_usage_t usage;
-  amdsmi_status_t status = amdsmi_get_gpu_activity_p(device_handles[event->device], &usage);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  switch (event->variant) {
-  case 0:
-    event->value = usage.gfx_activity;
-    break;
-  case 1:
-    event->value = usage.umc_activity;
-    break;
-  case 2:
-    event->value = usage.mm_activity;
-    break;
-  default:
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_fan_speed_max(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  int64_t max_speed = 0;
-  amdsmi_status_t status = amdsmi_get_gpu_fan_speed_max_p(device_handles[event->device], event->subvariant, &max_speed);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = max_speed;
-  return PAPI_OK;
-}
-static int access_amdsmi_pci_bandwidth(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_pcie_bandwidth_t bw;
-  amdsmi_status_t status = amdsmi_get_gpu_pci_bandwidth_p(device_handles[event->device], &bw);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  uint32_t cur_index = bw.transfer_rate.current;
-  if (cur_index >= bw.transfer_rate.num_supported) {
-    return PAPI_EMISC;
-  }
-  switch (event->variant) {
-  case 0:
-    event->value = bw.transfer_rate.num_supported;
-    break;
-  case 1:
-    event->value = (int64_t)bw.transfer_rate.frequency[cur_index];
-    break;
-  case 2:
-    event->value = bw.lanes[cur_index];
-    break;
-  default:
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_energy_count(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint64_t energy = 0;
-  float resolution = 0.0;
-  uint64_t timestamp = 0;
-  amdsmi_status_t status = amdsmi_get_energy_count_p(device_handles[event->device], &energy, &resolution, &timestamp);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  // Convert energy count to microJoules
-  event->value = (int64_t)(energy * resolution);
-  return PAPI_OK;
-}
-static int access_amdsmi_power_profile_status(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_power_profile_status_t status_info;
-  amdsmi_status_t status = amdsmi_get_gpu_power_profile_presets_p(device_handles[event->device], 0, &status_info);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  if (event->variant == 0) {
-    event->value = status_info.num_profiles;
-  } else if (event->variant == 1) {
-    event->value = (int64_t)status_info.current;
-  } else {
-    return PAPI_EMISC;
-  }
-  return PAPI_OK;
-}
-#ifndef AMDSMI_DISABLE_ESMI
-/* The functions below implement CPU metrics access */
-static int access_amdsmi_cpu_socket_power(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint32_t power = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_socket_power_p(device_handles[event->device], &power);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)power;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_socket_energy(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint64_t energy = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_socket_energy_p(device_handles[event->device], &energy);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)energy;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_socket_freq_limit(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint16_t freq = 0;
-  char *src = NULL;
-  amdsmi_status_t status = amdsmi_get_cpu_socket_current_active_freq_limit_p(device_handles[event->device], &freq, &src);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  if (src)
-    free(src);
-  event->value = freq;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_socket_freq_range(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint16_t fmax = 0, fmin = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_socket_freq_range_p(device_handles[event->device], &fmax, &fmin);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  if (event->variant == 0) {
-    event->value = fmin;
-  } else {
-    event->value = fmax;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_power_cap(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint32_t cap_value = 0;
-  amdsmi_status_t status;
-  if (event->variant == 0) {
-    status = amdsmi_get_cpu_socket_power_cap_p(device_handles[event->device], &cap_value);
-  } else {
-    status = amdsmi_get_cpu_socket_power_cap_max_p(device_handles[event->device], &cap_value);
-  }
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)cap_value;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_core_energy(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  int s_index = event->device - gpu_count;
-  if (s_index < 0 || s_index >= cpu_count) {
-    return PAPI_EMISC;
-  }
-  uint64_t energy = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_core_energy_p(cpu_core_handles[s_index][event->subvariant], &energy);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)energy;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_core_freq_limit(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  int s_index = event->device - gpu_count;
-  if (s_index < 0 || s_index >= cpu_count) {
-    return PAPI_EMISC;
-  }
-  uint32_t freq = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_core_current_freq_limit_p(cpu_core_handles[s_index][event->subvariant], &freq);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = freq;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_core_boostlimit(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  int s_index = event->device - gpu_count;
-  if (s_index < 0 || s_index >= cpu_count) {
-    return PAPI_EMISC;
-  }
-  uint32_t boost = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_core_boostlimit_p(cpu_core_handles[s_index][event->subvariant], &boost);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = boost;
-  return PAPI_OK;
-}
-static int access_amdsmi_dimm_temp(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_dimm_thermal_t dimm_temp;
-  amdsmi_status_t status = amdsmi_get_cpu_dimm_thermal_sensor_p(device_handles[event->device], (uint8_t)event->subvariant, &dimm_temp);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = (int64_t)dimm_temp.temp;
-  return PAPI_OK;
-}
-static int access_amdsmi_dimm_power(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_dimm_power_t dimm_pow;
-  amdsmi_status_t status = amdsmi_get_cpu_dimm_power_consumption_p(device_handles[event->device], (uint8_t)event->subvariant, &dimm_pow);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  event->value = dimm_pow.power;
-  return PAPI_OK;
-}
-static int access_amdsmi_dimm_range_refresh(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_temp_range_refresh_rate_t rate;
-  amdsmi_status_t status =
-      amdsmi_get_cpu_dimm_temp_range_and_refresh_rate_p(device_handles[event->device], (uint8_t)event->subvariant, &rate);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  if (event->variant == 0) {
-    event->value = rate.range;
-  } else {
-    event->value = rate.ref_rate;
-  }
-  return PAPI_OK;
-}
-static int access_amdsmi_threads_per_core(int mode, void *arg) {
-  (void)arg;
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint32_t threads = 0;
-  amdsmi_status_t status = amdsmi_get_threads_per_core_p(&threads);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  ((native_event_t *)arg)->value = threads;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_family(int mode, void *arg) {
-  (void)arg;
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint32_t family = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_family_p(&family);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  ((native_event_t *)arg)->value = family;
-  return PAPI_OK;
-}
-static int access_amdsmi_cpu_model(int mode, void *arg) {
-  (void)arg;
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  uint32_t model = 0;
-  amdsmi_status_t status = amdsmi_get_cpu_model_p(&model);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  ((native_event_t *)arg)->value = model;
-  return PAPI_OK;
-}
-static int access_amdsmi_smu_fw_version(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_smu_fw_version_t fw;
-  amdsmi_status_t status = amdsmi_get_cpu_smu_fw_version_p(device_handles[event->device], &fw);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  int encoded = ((int)fw.major << 16) | ((int)fw.minor << 8) | fw.debug;
-  event->value = encoded;
-  return PAPI_OK;
-}
-static int access_amdsmi_xgmi_bandwidth(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  amdsmi_processor_handle src = device_handles[event->device];
-  amdsmi_processor_handle dst = device_handles[gpu_count + event->subvariant];
-  uint64_t min_bw = 0, max_bw = 0;
-  amdsmi_status_t status = amdsmi_get_minmax_bandwidth_between_processors_p(src, dst, &min_bw, &max_bw);
-  if (status != AMDSMI_STATUS_SUCCESS) {
-    return PAPI_EMISC;
-  }
-  if (event->variant == 0) {
-    event->value = (int64_t)min_bw;
-  } else {
-    event->value = (int64_t)max_bw;
-  }
-  return PAPI_OK;
-}
-#endif
-
-static int access_amdsmi_cache_size(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) {
-    return PAPI_ENOSUPP;
-  }
-  if (!amdsmi_get_gpu_cache_info_p) return PAPI_ENOSUPP;
-
-  amdsmi_gpu_cache_info_t info;
-  amdsmi_status_t st = amdsmi_get_gpu_cache_info_p(device_handles[event->device], &info);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-
-  /* subvariant = cache index chosen during registration */
-  if (event->subvariant >= info.num_cache_types) return PAPI_EMISC;
-
-  /* cache_size is KB per header; expose bytes to stay consistent with other sizes */
-  uint64_t size_kb = info.cache[event->subvariant].cache_size;
-  event->value = (uint64_t)size_kb * 1024ULL;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_overdrive_level(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_overdrive_level_p) return PAPI_ENOSUPP;
-
-  uint32_t od = 0;
-  amdsmi_status_t st = amdsmi_get_gpu_overdrive_level_p(device_handles[event->device], &od);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)od;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_mem_overdrive_level(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_mem_overdrive_level_p) return PAPI_ENOSUPP;
-
-  uint32_t od = 0;
-  amdsmi_status_t st = amdsmi_get_gpu_mem_overdrive_level_p(device_handles[event->device], &od);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)od;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_od_volt_regions_count(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_od_volt_curve_regions_p) return PAPI_ENOSUPP;
-
-  /* Probe to get count; API requires a buffer, so do a two-call pattern */
-  uint32_t num = 0;
-  amdsmi_freq_volt_region_t *buf = NULL;
-
-  /* First call: ask for 0 (expect MORE_DATA/INSUFFICIENT_SIZE with num set) */
-  amdsmi_status_t st = amdsmi_get_gpu_od_volt_curve_regions_p(device_handles[event->device], &num, buf);
-  if (st == AMDSMI_STATUS_INSUFFICIENT_SIZE || st == AMDSMI_STATUS_NO_DATA) {
-    if (num == 0) return PAPI_EMISC;
-    buf = (amdsmi_freq_volt_region_t *)papi_calloc(num, sizeof(amdsmi_freq_volt_region_t));
-    if (!buf) return PAPI_ENOMEM;
-    st = amdsmi_get_gpu_od_volt_curve_regions_p(device_handles[event->device], &num, buf);
-  }
-  if (st != AMDSMI_STATUS_SUCCESS) {
-    if (buf) papi_free(buf);
-    return PAPI_EMISC;
-  }
-  event->value = (uint64_t)num;
-  if (buf) papi_free(buf);
-  return PAPI_OK;
-}
-
-static int access_amdsmi_perf_level(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_perf_level_p) return PAPI_ENOSUPP;
-
-  amdsmi_dev_perf_level_t perf = AMDSMI_DEV_PERF_LEVEL_UNKNOWN;
-  amdsmi_status_t st = amdsmi_get_gpu_perf_level_p(device_handles[event->device], &perf);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)perf;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_pm_metrics_count(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_pm_metrics_info_p) return PAPI_ENOSUPP;
-
-  amdsmi_name_value_t *metrics = NULL;
-  uint32_t count = 0;
-  amdsmi_status_t st = amdsmi_get_gpu_pm_metrics_info_p(device_handles[event->device], &metrics, &count);
-  if (metrics) papi_free(metrics); /* library allocates */
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)count;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_ras_ecc_schema(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_ras_feature_info_p) return PAPI_ENOSUPP;
-
-  amdsmi_ras_feature_t ras = {0};
-  amdsmi_status_t st = amdsmi_get_gpu_ras_feature_info_p(device_handles[event->device], &ras);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)ras.ecc_correction_schema_flag;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_reg_count(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_reg_table_info_p) return PAPI_ENOSUPP;
-
-  amdsmi_reg_type_t reg_type = (amdsmi_reg_type_t)event->variant; /* set at registration */
-  amdsmi_name_value_t *regs = NULL;
-  uint32_t num = 0;
-  amdsmi_status_t st = amdsmi_get_gpu_reg_table_info_p(device_handles[event->device], reg_type, &regs, &num);
-  if (regs) papi_free(regs);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)num;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_voltage(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_volt_metric_p) return PAPI_ENOSUPP;
-
-  amdsmi_voltage_type_t sensor = (amdsmi_voltage_type_t)event->subvariant; /* set at registration */
-  amdsmi_voltage_metric_t metric = (amdsmi_voltage_metric_t)event->variant; /* e.g., AMDSMI_VOLT_CURRENT */
-  int64_t mv = 0;
-  amdsmi_status_t st = amdsmi_get_gpu_volt_metric_p(device_handles[event->device], sensor, metric, &mv);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)mv; /* API reports mV */
-  return PAPI_OK;
-}
-
-static int access_amdsmi_vram_width(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_gpu_vram_info_p) return PAPI_ENOSUPP;
-
-  amdsmi_vram_info_t info;
-  amdsmi_status_t st = amdsmi_get_gpu_vram_info_p(device_handles[event->device], &info);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)info.vram_bit_width;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_soc_pstate_id(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_soc_pstate_p) return PAPI_ENOSUPP;
-
-  amdsmi_dpm_policy_t pol = {0};
-  amdsmi_status_t st = amdsmi_get_soc_pstate_p(device_handles[event->device], &pol);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)pol.current;
-  return PAPI_OK;
-}
-
-static int access_amdsmi_xgmi_plpd_id(int mode, void *arg) {
-  native_event_t *event = (native_event_t *)arg;
-  if (event->device < 0 || event->device >= device_count || !device_handles || !device_handles[event->device]) {
-    return PAPI_EMISC;
-  }
-  if (mode != PAPI_MODE_READ) return PAPI_ENOSUPP;
-  if (!amdsmi_get_xgmi_plpd_p) return PAPI_ENOSUPP;
-
-  amdsmi_dpm_policy_t pol = {0};
-  amdsmi_status_t st = amdsmi_get_xgmi_plpd_p(device_handles[event->device], &pol);
-  if (st != AMDSMI_STATUS_SUCCESS) return PAPI_EMISC;
-  event->value = (uint64_t)pol.current;
   return PAPI_OK;
 }
